assignment2/requirements.txt
matplotlib
numpy
pandas
Pillow
skimage
scikit-learn
torch
tqdm